--- src/cruds/Notification_alluser.php
@@ -1,27 +1,0 @@
-<?php
-
-namespace cruds;
-
-require_once('User.php');
-use cruds\User;
-
-class Notification_alluser
-{
-  public function __construct(\PDO $db)
-  {
-    $this->db = $db;
-  }
-
-  public function get_all_user() {
-    $stmt = $this->db ->prepare("SELECT email FROM users");
-    $stmt -> execute();
-    return $stmt -> fetchAll();
-  }
-
-  public function before_attendance_event() {
-    $stmt = $this->db ->prepare("SELECT name,detail,start_at,end_at FROM events WHERE DATE(start_at) = DATE_ADD(CURRENT_DATE, INTERVAL 1 DAY)");
-    $stmt -> execute();
-    return $stmt -> fetchAll();
-  }
- 
-}
